review docs of 'parse' module; refactor (#331)
