+++ git-config/src/parse/error.rs
@@ -1,8 +1,7 @@
 use crate::parse::Error;
 use std::fmt::Display;
 
-/// A list of parsers that parsing can fail on. This is used for pretty-printing
-/// errors
+/// A list of parsers that parsing can fail on. This is used for pretty-printing errors
 #[derive(PartialEq, Debug, Clone, Copy)]
 pub(crate) enum ParseNode {
     SectionHeader,
@@ -28,7 +27,7 @@ impl Error {
         self.line_number + 1
     }
 
-    /// The remaining data that was left unparsed.
+    /// The data that was left unparsed, which contains the cause of the parse error.
     #[must_use]
     pub fn remaining_data(&self) -> &[u8] {
         &self.parsed_until
@@ -37,8 +36,6 @@ impl Error {
 
 impl Display for Error {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        let data_size = self.parsed_until.len();
-        let data = std::str::from_utf8(&self.parsed_until);
         write!(
             f,
             "Got an unexpected token on line {} while trying to parse a {}: ",
@@ -46,18 +43,19 @@ impl Display for Error {
             self.last_attempted_parser,
         )?;
 
+        let data_size = self.parsed_until.len();
+        let data = std::str::from_utf8(&self.parsed_until);
         match (data, data_size) {
             (Ok(data), _) if data_size > 10 => {
-                write!(f, "'{}' ... ({} characters omitted)", &data[..10], data_size - 10)
+                write!(
+                    f,
+                    "'{}' ... ({} characters omitted)",
+                    &data.chars().take(10).collect::<String>(),
+                    data_size - 10
+                )
             }
             (Ok(data), _) => write!(f, "'{}'", data),
-            (Err(_), _) if data_size > 10 => write!(
-                f,
-                "'{:02x?}' ... ({} characters omitted)",
-                &self.parsed_until[..10],
-                data_size - 10
-            ),
-            (Err(_), _) => write!(f, "'{:02x?}'", self.parsed_until),
+            (Err(_), _) => self.parsed_until.fmt(f),
         }
     }
 }
