+++ git-config/src/parse/tests.rs
@@ -1,56 +1,3 @@
-mod parse {
-    use crate::parse::Events;
-
-    #[test]
-    fn parser_skips_bom() {
-        let bytes = b"
-        [core]
-            a = 1
-    ";
-        let bytes_with_gb18030_bom = "\u{feff}
-        [core]
-            a = 1
-    ";
-
-        assert_eq!(
-            Events::from_bytes(bytes),
-            Events::from_bytes(bytes_with_gb18030_bom.as_bytes())
-        );
-        assert_eq!(
-            Events::from_bytes_owned(bytes, None),
-            Events::from_bytes_owned(bytes_with_gb18030_bom.as_bytes(), None)
-        );
-    }
-}
-
-#[cfg(test)]
-mod error {
-    use crate::parse::Events;
-
-    #[test]
-    fn line_no_is_one_indexed() {
-        assert_eq!(Events::from_str("[hello").unwrap_err().line_number(), 1);
-    }
-
-    #[test]
-    fn remaining_data_contains_bad_tokens() {
-        assert_eq!(Events::from_str("[hello").unwrap_err().remaining_data(), b"[hello");
-    }
-
-    #[test]
-    fn to_string_truncates_extra_values() {
-        assert_eq!(
-            Events::from_str("[1234567890").unwrap_err().to_string(),
-            "Got an unexpected token on line 1 while trying to parse a section header: '[123456789' ... (1 characters omitted)"
-        );
-    }
-
-    #[test]
-    fn detected_by_fuzz() {
-        assert!(Events::from_str("[]I=").is_err());
-    }
-}
-
 pub(crate) mod util {
     //! This module is only included for tests, and contains common unit test helper
     //! functions.
